fix: fix chaining issue & major issue with the request limiters thread safety (#25)
request limiter had a major flaw.. first of all... it didnt matter which key got used.. i didnt make a difference when thinking about queued requests. The key specific lock should handle that. Now there is also no need to lock and unlock the thread again and removing the timestamp and so on...